requirements.txt
# requirements.txt

backoff==2.2.1
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coinbase-advanced-py==1.3.0
cryptography==42.0.7
dacite==1.8.1
idna==3.7
iniconfig==2.0.0
mypy==1.10.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
pytest==8.2.0
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
